requirements.txt
anyio==3.5.0
appnope==0.1.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
async-timeout==4.0.2
attrs==21.4.0
Babel==2.9.1
backcall==0.2.0
bcrypt==4.0.1
beautifulsoup4==4.11.1
bleach==4.1.0
brotlipy==0.7.0
cachelib==0.9.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cryptography==38.0.1
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.13
distlib==0.3.6
docopt==0.6.2
entrypoints==0.4
executing==0.8.3
fastjsonschema==2.16.2
filelock==3.9.0
Flask==2.2.2
Flask-Bcrypt==1.0.1
Flask-Cors==3.0.10
Flask-Session==0.4.0
Flask-SQLAlchemy==3.0.2
Flask-WTF==1.0.1
greenlet==2.0.1
idna==3.4
ipykernel==6.15.2
ipython==8.4.0
ipython-genutils==0.2.0
ipywidgets==7.6.5
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
json5==0.9.6
jsonschema==4.16.0
jupyter==1.0.0
jupyter_client==7.3.5
jupyter-console==6.4.3
jupyter_core==4.11.1
jupyter-server==1.18.1
jupyterlab==3.4.4
jupyterlab-pygments==0.1.2
jupyterlab_server==2.15.2
jupyterlab-widgets==1.0.0
MarkupSafe==2.1.1
marshmallow==3.19.0
matplotlib-inline==0.1.6
mistune==0.8.4
nbclassic==0.3.5
nbclient==0.5.13
nbconvert==6.4.4
nbformat==5.5.0
nest-asyncio==1.5.5
notebook==6.4.12
packaging==21.3
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
pip==22.2.2
pipenv==2022.12.19
pipreqs==0.4.11
platformdirs==2.6.2
ply==3.11
prometheus-client==0.14.1
prompt-toolkit==3.0.20
psutil==5.9.0
psycopg2-binary==2.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
Pygments==2.11.2
pyOpenSSL==22.0.0
pyparsing==3.0.9
PyQt5-sip==12.11.0
pyrsistent==0.18.0
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.1
pyzmq==23.2.0
qtconsole==5.3.2
QtPy==2.2.0
redis==4.3.4
requests==2.28.1
Send2Trash==1.8.0
setuptools==63.4.1
sip==6.6.2
six==1.16.0
sniffio==1.2.0
soupsieve==2.3.2.post1
spotipy==2.21.0
SQLAlchemy==1.4.44
stack-data==0.2.0
terminado==0.13.1
testpath==0.6.0
toml==0.10.2
tornado==6.2
traitlets==5.1.1
typing_extensions==4.3.0
urllib3==1.26.12
virtualenv==20.17.1
virtualenv-clone==0.5.7
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==0.58.0
Werkzeug==2.2.2
wheel==0.37.1
widgetsnbextension==3.5.2
wrapt==1.14.1
WTForms==3.0.1
yarg==0.1.9
